add chat-hub; add DTOs, service, controller of user; refact: models

+++ ChatApp.DOMAIN/Services/Chat/ChatService.cs
@@ -17,41 +17,32 @@ namespace ChatApp.DOMAIN
             var chats = await _context.Chats.ToListAsync();
 
             if (chats == null)
-                throw new CustomException(CustomExceptionType.NotFound,
-                    "Chats weren't found.");
+                throw new CustomException(CustomExceptionType.NotFound, "Chats weren't found.");
 
-            var chatDTOs = new List<ChatDTO>();
-
-            foreach (var chat in chats)
-            {
-                var chatDTO = ChatDTO.ChatToChatDTO(chat);
-                chatDTOs.Add(chatDTO);
-            }
+            var chatDTOs = chats.Select(ChatDTO.ChatToChatDTO).ToList();
 
             return chatDTOs;
         }
 
-        public async Task<ChatDTO> GetChatByIdAsync(int id)
+        public async Task<ChatByIdDTO> GetChatByIdAsync(int id)
         {
-            var chat = await _context.Chats.FindAsync(id);
+            var chat = await _context.Chats.Include(c => c.Messages).ThenInclude(m => m.User)
+                .FirstOrDefaultAsync(c => c.ChatId == id);
 
             if (chat == null)
-                throw new CustomException(CustomExceptionType.NotFound,
-                    $"Chat id {id} wasn't found.");
+                throw new CustomException(CustomExceptionType.NotFound, $"Chat id {id} wasn't found.");
 
-            var chatDTO = ChatDTO.ChatToChatDTO(chat);
+            var chatDTO = ChatByIdDTO.ChatToChatByIdDTO(chat);
 
             return chatDTO;
         }
 
-        public async Task<ChatDTO> CreateChatAsync(CreateChatDTO request)
+        public async Task<ChatDTO> CreateChatAsync(CreateChatDTO request, int userId)
         {
-            var chatByFullName = await _context.Chats
-                .FirstOrDefaultAsync(c => c.ChatName == request.ChatName);
+            var chatByFullName = await _context.Chats.FirstOrDefaultAsync(c => c.ChatName == request.ChatName);
 
             if (chatByFullName != null)
-                throw new CustomException(CustomExceptionType.ChatIsAlreadyExists,
-                    $"Chat is already exists.");
+                throw new CustomException(CustomExceptionType.ChatIsAlreadyExists, "Chat already exists.");
 
             var chat = CreateChatDTO.CreateChatDTOToChat(request);
 
@@ -69,8 +60,7 @@ namespace ChatApp.DOMAIN
             var chatById = await _context.Chats.FindAsync(id);
 
             if (chatById == null)
-                throw new CustomException(CustomExceptionType.NotFound,
-                    $"Chat id {id} wasn't found.");
+                throw new CustomException(CustomExceptionType.NotFound, $"Chat id {id} wasn't found.");
 
             UpdateChatDTO.UpdateChat(chatById, request);
 
@@ -82,13 +72,15 @@ namespace ChatApp.DOMAIN
             return chatDTO;
         }
 
-        public async Task DeleteChatAsync(int id)
+        public async Task DeleteChatAsync(int id, int userId)
         {
             var chatById = await _context.Chats.FindAsync(id);
 
             if (chatById == null)
-                throw new CustomException(CustomExceptionType.NotFound,
-                    $"Chat id {id} wasn't found.");
+                throw new CustomException(CustomExceptionType.NotFound, $"Chat id {id} wasn't found.");
+
+            if (chatById.CreatorUserId != userId)
+                throw new CustomException(CustomExceptionType.NoPermissions, "No permissions to perform the operation.");
 
             _context.Chats.Remove(chatById);
             await _context.SaveChangesAsync();
